+++ CHANGELOG.md
@@ -2,3 +2,5 @@
 ==========
 
 * When filling a text column, replace `null` with empty text.
+* Support timestamp columns.
+* Support filling a column with values from other columns (like SQL COALESCE()")